+++ .storybook/preview.ts
@@ -22,7 +22,7 @@ const preview: Preview = {
     },
     (story, { globals: { mode = 'light', theme = 'sanoma-learning' } }) => {
       updateTheme(theme, mode);
-      
+
       return story();
     }
   ],
@@ -44,7 +44,7 @@ const preview: Preview = {
         dynamicTitle: true,
         icon: 'mirror',
         items: [
-          { value: 'light', left: '🌞', title: 'Light mode' }, 
+          { value: 'light', left: '🌞', title: 'Light mode' },
           { value: 'dark', left: '🌛', title: 'Dark mode' },
         ],
       }