+++ packages/components/form/index.ts
@@ -1,4 +1,6 @@
 export * from './src/error.js';
 export * from './src/form-control-mixin.js';
+export * from './src/form-field.js';
 export * from './src/hint.js';
 export * from './src/label.js';
+export * from './src/update-validity-event.js';